+++ tdmclient/tools/list.py
@@ -26,12 +26,15 @@ Options:
   --tdmport=P    tdm port (default: 8596 (tcp) or 8597 (ws), or from zeroconf)
   --tdmws        connect to tdm with WebSocket (default: plain TCP)
   --zeroconf     use zeroconf (default: no zeroconf)
+  --zcall        discover TDM information published on all interfaces instead
+                 of only default one
 """)
 
 
 def main(argv=None, tdm_transport=None):
     debug = 0
     zeroconf = False
+    zeroconf_all = False
     tdm_addr = None
     tdm_port = None
     tdm_ws = False
@@ -53,6 +56,7 @@ def main(argv=None, tdm_transport=None):
                                                   "tdmport=",
                                                   "tdmws",
                                                   "zeroconf",
+                                                  "zcall",
                                               ])
         except getopt.error as err:
             print(str(err))
@@ -77,8 +81,11 @@ def main(argv=None, tdm_transport=None):
                 tdm_ws = True
             elif arg == "--zeroconf":
                 zeroconf = True
+            elif arg == "--zcall":
+                zeroconf = True
+                zeroconf_all = True
 
-    with ClientAsync(zeroconf=zeroconf,
+    with ClientAsync(zeroconf=zeroconf, zeroconf_all=zeroconf_all,
                      tdm_addr=tdm_addr, tdm_port=tdm_port, tdm_ws=tdm_ws,
                      tdm_transport=tdm_transport,
                      password=password,