+++ tdmclient/tools/run.py
@@ -39,6 +39,8 @@ Options:
   --tdmport=P    tdm port (default: 8596 (tcp) or 8597 (ws), or from zeroconf)
   --tdmws        connect to tdm with WebSocket (default: plain TCP)
   --zeroconf     use zeroconf (default: no zeroconf)
+  --zcall        discover TDM information published on all interfaces instead
+                 of only default one
 """, **kwargs)
 
 
@@ -48,6 +50,7 @@ def main(argv=None, tdm_transport=None):
     stop = False
     scratchpad = 0  # 1=--scratchpad, 2=--sponly
     zeroconf = False
+    zeroconf_all = False
     tdm_addr = None
     tdm_port = None
     tdm_ws = False
@@ -103,6 +106,7 @@ def main(argv=None, tdm_transport=None):
                                                   "tdmaddr=",
                                                   "tdmport=",
                                                   "tdmws",
+                                                  "zcall",
                                                   "zeroconf",
                                               ])
         except getopt.error as err:
@@ -151,6 +155,9 @@ def main(argv=None, tdm_transport=None):
                 tdm_ws = True
             elif arg == "--zeroconf":
                 zeroconf = True
+            elif arg == "--zcall":
+                zeroconf = True
+                zeroconf_all = True
 
     if stop:
         if len(values) > 0:
@@ -209,7 +216,7 @@ def main(argv=None, tdm_transport=None):
         for event_name in transpiler.events_out:
             events.append((event_name, transpiler.events_out[event_name]))
 
-    with ClientAsync(zeroconf=zeroconf,
+    with ClientAsync(zeroconf=zeroconf, zeroconf_all=zeroconf_all,
                      tdm_addr=tdm_addr, tdm_port=tdm_port, tdm_ws=tdm_ws,
                      tdm_transport=tdm_transport,
                      password=password,